+++ Emulator/Files/AnyCollection.h
@@ -26,13 +26,13 @@ public:
     virtual u64 collectionCount() const = 0;
     
     // Returns the name of a certain item
-    virtual PETName<16> itemName(unsigned nr) const = 0;
+    virtual PETName<16> itemName(isize nr) const = 0;
     
     // Returns the size of a certain in bytes
-    virtual u64 itemSize(unsigned nr) const = 0;
+    virtual u64 itemSize(isize nr) const = 0;
         
     // Reads a byte from a certain item
-    virtual u8 readByte(unsigned nr, u64 pos) const = 0;
+    virtual u8 readByte(isize nr, u64 pos) const = 0;
 
     
     //
@@ -40,12 +40,12 @@ public:
     //
 
     // Reads a word from a certain item in little endian or big endian format
-    u16 readWordBE(unsigned nr, u64 pos) const;
-    u16 readWordLE(unsigned nr, u64 pos) const;
+    u16 readWordBE(isize nr, u64 pos) const;
+    u16 readWordLE(isize nr, u64 pos) const;
     
     // Return the load address of this item
-    u16 itemLoadAddr(unsigned nr) const;
+    u16 itemLoadAddr(isize nr) const;
 
     // Copies an item into a buffer
-    virtual void copyItem(unsigned nr, u8 *buf, u64 len, u64 offset = 0) const;
+    virtual void copyItem(isize nr, u8 *buf, u64 len, u64 offset = 0) const;
 };
